refactor: create auth/app routes
