+++ web/src/App.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react';
 
 import './assets/styles/global.css';
 
-import Routes from './routes';
+import Routes from './routes/index';
 
 const App: FC = () => {
   return <Routes />;
