+++ web/src/routes/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { useAuth } from '../contexts/auth';
+
+import AppRoutes from './app.routes';
+import AuthRoutes from './auth.routes';
+
+const Routes: React.FC = () => {
+  const { signed } = useAuth();
+
+  return signed ? <AppRoutes /> : <AuthRoutes />;
+};
+
+export default Routes;
